feat: move renderer code from game_ui to lib, renaming to Velyst
